Prevent 'p' parameter of OPTIONS requests being logged.
Previously, OPTIONS requests were logged without passwords being redacted.

This occurred because password redaction is performed by the authentication
handler and a) the serveOptions handler did not have the parameter
that forces the authentication handler to be put on the request path and b)
the CORS handler returns early for OPTIONS requests
before the authentication handler is invoked.

With this commit, OPTIONS requests are also subject to the authentication
and the responsibility for the early return on OPTIONS request is moved from
the CORS handler to the authentication handler.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>